Refactoring: Create new facade class and move some static methods from MailQue into the new one.
