+++ src/main/java/org/masukomi/aspirin/core/MailQue.java
@@ -33,7 +33,6 @@ import java.util.Iterator;
 import java.util.Vector;
 
 import javax.mail.MessagingException;
-import javax.mail.Session;
 import javax.mail.internet.AddressException;
 import javax.mail.internet.MimeMessage;
 
@@ -57,11 +56,6 @@ public class MailQue implements MailQueMBean {
 	private Vector<AspirinListener> listenersToAdd;
 	private int notificationCount;
 	
-	/** This session is used to generate new MimeMessage objects. */
-	private static Session defaultSession = null;
-	/** This counter is used to generate unique message ids. */
-	private static Integer idCounter = 0;
-	
 	public MailQue() {
 		qm = new QueManager(this);
 		que = new Vector<QuedItem>();
@@ -77,23 +71,11 @@ public class MailQue implements MailQueMBean {
 		notifyQueManager();
 	}
 	
-	private String getAspirinMailID(MimeMessage message) {
-		String[] headers;
-		try {
-			headers = message.getHeader(Configuration.ASPIRIN_MAIL_ID_HEADER);
-			if( headers != null && 0 < headers.length )
-				return headers[0];
-		} catch (MessagingException e) {
-			Configuration.getInstance().getLog().error("Header could not be get from MimeMessage.", e);
-		}
-		return message.toString();
-	}
-	
 	protected void service(MimeMessage mimeMessage, Collection<AspirinListener> watchers)
 			throws AddressException, MessagingException {
 
 		MailImpl sourceMail = new MailImpl(mimeMessage);
-		sourceMail.setName(getAspirinMailID(mimeMessage));
+		sourceMail.setName(Aspirin.getMailID(mimeMessage));
 		// Do I want to give the internal key, or the message's Message ID
 		if (log.isDebugEnabled())
 			log.debug(getClass().getSimpleName()+".service(): Remotely delivering mail " + sourceMail.getName());
@@ -262,23 +244,6 @@ public class MailQue implements MailQueMBean {
 			getQueManager().start();
 		else
 			getQueManager().notifyWithMail();
-	}
-	
-	public static MimeMessage createNewMimeMessage() {
-		if( defaultSession == null )
-			defaultSession = Session.getDefaultInstance(System.getProperties());
-		MimeMessage mMesg = new MimeMessage(defaultSession);
-		synchronized (idCounter) {
-			long nowTime = System.currentTimeMillis()/1000;
-			String newId = nowTime+"."+Integer.toHexString(idCounter++);
-			try {
-				mMesg.setHeader(Configuration.ASPIRIN_MAIL_ID_HEADER, newId);
-			} catch (MessagingException msge) {
-				Configuration.getInstance().getLog().warn("Aspirin Mail ID could not be generated.", msge);
-				msge.printStackTrace();
-			}
-		}
-		return mMesg;
 	}
 
 }
